requirements.txt
pillow>=9.0.0
colorama>=0.4.0
gradio>=4.0.0
